convert to long like in ae2 code

+++ src/main/java/com/glodblock/github/client/gui/GuiPatternMulti.java
@@ -115,7 +115,7 @@ public class GuiPatternMulti extends GuiAmount {
     }
 
     @Override
-    protected int addOrderAmount(final int i) {
+    protected long addOrderAmount(final int i) {
         return i + getAmount();
     }
